Simplify browser includes
